requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
idna==3.11
importlib_resources==6.5.2
joblib==1.5.3
kafka-python==2.3.0
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pyarrow==21.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3
websocket-client==1.9.0
zipp==3.23.0
